Kernels are building but not running on troi

+++ src/cl_kernels/useDouble.cl
@@ -1,7 +1,10 @@
 #ifndef USE_DOUBLE_CL_
 #define USE_DOUBLE_CL_
 
-typedef float FLOAT;
+//typedef float FLOAT;
+//typedef float4 FLOAT4;
+
+typedef double FLOAT; 
 typedef float4 FLOAT4;
 
 #endif
